backend/requirements.txt
fastapi
uvicorn
pipecat
google-generativeai
websockets
python-dotenv